requirements.txt
Django==2.0
django-filter==1.1.0
djangorestframework==3.7.3
Markdown==2.6.9
olefile==0.44
Pillow==4.3.0
pytz==2017.3
